info/synapse-map-tables-new/requirements.txt
azure-functions>=1.17.0
azure-storage-blob==12.24.0
pymssql
pyarrow
